added transparency to the hot & cold zones
Modified program so that the graphic now illustrates how much better or worse a player is shooting compared to the league average. This was accomplished by modifying the transparency of the matplotlib elements for each shot zone so that it matches the difference between a player's shooting stats and the league average stats.